+++ ThirdPartyAdapters/tapjoy/tapjoy/src/main/java/com/google/ads/mediation/tapjoy/TapjoyAdapter.java
@@ -469,6 +469,7 @@ public class TapjoyAdapter implements MediationInterstitialAdapter, MediationRew
         mainHandler.post(new Runnable() {
             @Override
             public void run() {
+                mediationRewardedVideoAdListener.onVideoCompleted(TapjoyAdapter.this);
                 mediationRewardedVideoAdListener.onRewarded(TapjoyAdapter.this, new TapjoyReward());
             }
         });