+++ ThirdPartyAdapters/vungle/vungle/src/main/java/com/vungle/mediation/VungleListener.java
@@ -20,24 +20,19 @@ abstract class VungleListener {
         return mIsWaitingInit;
     }
 
-
     void setWaitingInit(boolean isWaitingInit) {
         this.mIsWaitingInit = isWaitingInit;
     }
 
-    void onAdEnd(String placement, boolean wasSuccessfulView, boolean wasCallToActionClicked) {
+    void onAdEnd(String placement, boolean wasSuccessfulView, boolean wasCallToActionClicked) {}
 
-    }
+    void onAdStart(String placement) {}
 
-    void onAdStart(String placement) {
-    }
+    void onAdFail(String placement) {}
 
-    void onAdFail(String placement) {
-    }
+    void onAdAvailable() {}
 
-    void onAdAvailable() {
-    }
+    void onAdFailedToLoad() {}
 
-    void onInitialized(boolean isSuccess) {
-    }
+    void onInitialized(boolean isSuccess) {}
 }